Cache: Add caching of base results and snapshots
After running experiments on base branch, the results
are cached to the host (`.cache` dir), also the snapshots are cached.
The results are used in subsequent runs so the comparison does not
have to be redone on base branch. The cached snapshots are used
when running experiments on pr branch, they are recovered and only
compared which should be enough in most cases.