Add AbortController for aborting evaluation
The evaluation now allows to abort itself by triggering abort signal on
abortController.
When the `abort` event is emitted containers are killed,
if some command are run in the containers, the container should
throw error. The error is caught and check if the abort signal is signal
and if yes, then abort signal is rethrown - the error should end
the evaluation.
If command was not run in a container, there are checks in some
places (mainly when leaving evaluation methods)
which check if abort signal was emitted and if yes
it throws error, therefore the evaluation will be aborted.

Also added tests for testing abortion mechanism.