Add script for evaluating past (closed) PRs
The script evaluates the PR according to the base branch
at that time, reports differences to README files and
saves results to directory.

The script is able to handle analyzing of PRs that need to have
specified PR comparison options (e.g. for the SMT extension).

When analyzing PRs with a script, saving info about the PR to the
report file. The following info is saved:
- base repo name and SHA,
- PR repo name and SHA,
- verdict of the analysis (same/differnt results),
- errors.