DiffKemp: Run development shell durring setup
Running the shell during setup method of DiffKemp so
the necessary dependencies for development shell are
downloaded before the shell will be used. The reason is
that otherwise the shell could be used for the first time
used in parallel / multiple times at the same time
(e.g. when using build-kernel). This commit
is created to avoid problems connected with it.