Cache: Restoring snapshots using PR SHA
This commit allows to specify multiple commit SHAs which
should be checked when recovering snapshots for PR comparison.
This allows to also specify PR SHA.
When recovering snapshots for PR, the SHAs (identification
on which DiffKemp commit the snapshot was build) are checked in
order -- firstly PR snapshots, then base snapshots, if neither of
them exists they are created.

Note: If snapshots were recovered from base branch and PR snapshots
caching is enabled, they will be saved also as PR snapshots (which
can be unexpected behavior).