RHEL: Extract fun definitions for RHEL sysctl
For RHEL sysctl the definitions of functions were not extracted
because it is not yet correctly implemented in DiffKemp.
For this reason, it was not possible to show functions diffs
for RHEL sysctl in PR reports.
Solving this by creating directory with patches that are aplicated
on DiffKemp. This commit also adds a patch that fixes the extraction
respectively saving of functions definitions to diffkemp-out.yaml file
for sysctl comparison.