Fix building of Nix flakes for older DiffKemp
It was not possible to build DiffKemp which used llvm16
on the server before. Now it should be possible, because
the added patch turns off testing of the nix build.
The testing previously caused failure of the build because
of a circular dependency.

The patch is applicated differently than the previous patch
(.diffkemp-patches). Patches in (.diffkemp-patches) are mounted
to container at runtime and are applicated when settuping DiffKemp.
Because this patch is necessary when building the DiffKemp
(without it DiffKemp would not be able to build) and because
DiffKemp is pre-build in Dockerfile, the patch is not mounted but
copied to the container so it can be pre-build.
Note: It is necessary to rebuild the container image when adding
patches to the `.build-patches` directory.