Fix cloning private repository for development
A token necessary for cloning private repositories (used for
testing/development) was previously created when the evaluation
was initiated. The token is valid only for a limited time and in
case the evaluation was enqueued because another evaluation was
running, the token expired and the private repository could not
be cloned. This commit fixes it by creating the token after the
evaluation is dequeued. This required changes in `EvaluationConfig`.