Container: Split preparation of RHEL to two layers
Downloading RHEL kernels for function comparison and
the copying them was in the same layer. This caused problems
while building the container image, because creating the layer
took (1) too long and (2) lot of a disk space.
This commits split the downloading of RHEL kernels to first layer
and copying them for sysctl parameter comparison to second layer,
the reason is to reduce the disk space required for building
the container image. It looks like this helped.