Cache: Do not cache snapshots for closed PRs
By default caching only of the last snapshot is enabled, so the
we do not unnecessary cache snapshots for multiple commits which
would fill up the disk quickly.
When evaluating past PRs, the old base branch would be cached and it
would replaced the newer base branch snapshot.
This is not very useful, for that reason, this commit adds check
-- caching the base snapshot only if it is newer than the already
cached one.