Handle updates of PR branches from forked repos
The bot have handled aborting of running evaluations if the PR
branch with running evaluation was updated. It only handled PR branches
created on the base repository (on the repository where the bot is
installed). If the branch was located on a fork of the repository
and the branch was updated, the bot did not aborted running evaluations.
This PR fixes it by adding support for `pull_request.synchronize`
webhook which is triggered when the PR branch (located in different
fork/repo) is updated. When this events is received running evaluations
for the branch are aborted.