Remove evaluation labels on push to PR
When the PR is changed by push, we can expect that
the push changed the code and theoretically it could change
the results of evaluation. For this reason, if push to PR
happens we are now removing all evaluation labels from the PR.
The PR reviewer should rerun the evaluation if he wants to known
the current results.